services/task-engine/pyproject.toml
[project]
name = "task-engine"
version = "0.1.0"
description = "LLM-powered task extraction service for Kutana AI"
requires-python = ">=3.12"
dependencies = [
    "kutana-core",
    "kutana-providers",
    "kutana-memory",
    "redis>=5.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "pydantic-settings>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.32",
]

[tool.uv.sources]
kutana-core = { workspace = true }
kutana-providers = { workspace = true }
kutana-memory = { workspace = true }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/task_engine"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
